scaling in the model, is not working

+++ 2L_SDVRP/include/LoadingChecker.h
@@ -57,10 +57,10 @@ class LoadingChecker
                                 const std::vector<Cuboid>   &items,
                                 double                       maxRuntimeSec,
                                 std::vector<Array<int, 4>>  &vetPos,
-                                int& fk,
-                                int& fFA,
-                                int& fRA,
-                                int& fTA);
+                                double& fk,
+                                double& fFA,
+                                double& fRA,
+                                double& fTA);
     // bool isCallTypeExact,
     // double maxRuntime = std::numeric_limits<double>::max());
