+++ 2L_SDVRP/src/c_api.cpp
@@ -235,7 +235,7 @@ int testRoute(int *vet_c, int vetSize, int onlyHeuristic, int doInverseRoute)
     //std::printf("\n\n************************************\n");
     //std::printf(    "**************INI CP-Sat************\n\n");
 
-    int int_fk, int_fFA, int_fRA, int_fTA;
+    double int_fk, int_fFA, int_fRA, int_fTA;
 
     for(int i=0; i < 1; ++i)
     {
